feat: Fixed makefile, added time functionality

+++ pkg/rest_api/models/response/timeresponse.go
@@ -2,5 +2,5 @@ package response
 
 type TimeResponse struct {
 	UTC string  `json:"utc"`
-	Unix int	`json:"unix"`
+	Unix int64	`json:"unix"`
 }
